Refactor OCS facility and forms from LCO ones
